IRL_requirements.txt
alabaster==0.7.10
appnope==0.1.0
argon2-cffi==16.3.0
attrs==17.4.0
awesome-slugify==1.6.5
Babel==2.5.3
backports.shutil-get-terminal-size==1.0.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
configparser==3.5.0
contextlib2==0.5.5
coverage==4.4.2
decorator==4.2.1
dj-database-url==0.5.0
Django==1.10.8
django-activity-stream==0.6.5
django-allauth==0.34.0
django-appconf==1.0.2
django-compressor==2.2
django-coverage-plugin==1.5.0
django-crispy-forms==1.7.0
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-extensions==1.9.8
django-markdownx==2.0.22
django-model-utils==3.0.0
django-redis==4.8.0
django-test-plus==1.0.20
docutils==0.14
enum34==1.1.6
factory-boy==2.9.2
Faker==0.8.12
flake8==3.5.0
funcsigs==1.0.2
gunicorn==19.7.1
idna==2.6
imagesize==1.0.0
ipaddress==1.0.19
ipdb==0.10.3
ipython==5.5.0
ipython-genutils==0.2.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.0.6
olefile==0.45.1
pathlib2==2.3.0
pexpect==4.4.0
pickleshare==0.7.4
Pillow==4.3.0
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
py==1.5.2
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
Pygments==2.2.0
pytest==3.4.2
pytest-django==3.1.2
pytest-sugar==0.9.0
python-dateutil==2.7.0
python-openid==2.2.5
pytz==2017.3
raven==6.6.0
rcssmin==1.0.6
redis==2.10.6
regex==2018.2.21
requests==2.18.4
requests-oauthlib==0.8.0
rjsmin==1.0.12
scandir==1.7
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.6.5
sphinxcontrib-websupport==1.0.1
sqlparse==0.2.4
termcolor==1.1.0
text-unidecode==1.2
traitlets==4.3.2
typing==3.6.4
Unidecode==0.4.21
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.13
whitenoise==3.3.1
